frontend

+++ admin/php/controllers/esquema.controller.php
@@ -4,7 +4,7 @@ include_once(INC.'php/bootstrap.php');
 
 if ($_POST['id']) {
     $esquema = Doctrine::getTable('esquema')->find($_POST['id']);
-    Doctrine_Query::create()->delete('Imagen')->where('id_esquema = ? ', $_POST['id'])->andWhereNotIn('id', $_POST['bloquesId'])->execute();
+    if (isset($_POST['bloquesId'])) Doctrine_Query::create()->delete('Imagen')->where('id_esquema = ? ', $_POST['id'])->andWhereNotIn('id', $_POST['bloquesId'])->execute();
 } else {
     $esquema = new Esquema();
 }
@@ -17,8 +17,8 @@ if (isset($_POST['bloquesId'])) {
     for ($i=0, $l= count($_POST['bloquesId']); $i<$l; $i++) {
         if ($_POST['bloquesId'][$i]!=0){
             $bloque = Doctrine::getTable('imagen')->find($_POST['bloquesId'][$i]);
-            @rename(INC.'../content/tmp/esquemas/'.$bloque->src, INC.'../content/esquemas/'.$bloque->src);
-            @rename(INC.'../content/tmp/esquemas/thumb/'.$bloque->src, INC.'../content/esquemas/thumb/'.$bloque->src);
+            rename(INC.'../content/tmp/esquemas/'.$bloque->src, INC.'../content/esquemas/'.$bloque->src);
+            rename(INC.'../content/tmp/esquemas/thumb/'.$bloque->src, INC.'../content/esquemas/thumb/'.$bloque->src);
         } else {
             $bloque = new Imagen();
         }
@@ -28,6 +28,7 @@ if (isset($_POST['bloquesId'])) {
         $esquema->bloques[] = $bloque;
     }
 }
+
 
 //imagen
 foreach (glob(INC.'../content/tmp/esquemas/'.$_POST['imgprefix'].'-*') as $img) {
@@ -40,8 +41,6 @@ foreach (glob(INC.'../content/tmp/esquemas/'.$_POST['imgprefix'].'-*') as $img) 
 
 $esquema->save();
 
-if (isset($_POST['redirect'])) {
-    $accion = ($_POST['id'])?'#edit':'#new';
-    //header('location: '.URL.'esquema'.$accion);
-}
+$accion = ($_POST['id'])?'#edit':'#new';
+header('location: '.URL.'regiones'.$accion);
 ?>
